Move test data to files.
This feels a bit easier to read than embedded strings.
