+++ gen/nodejs/generator_test.go
@@ -3,15 +3,12 @@ package nodejs
 import (
 	"bytes"
 	"io/ioutil"
-	"os"
-	"path"
 	"testing"
 
 	"github.com/hashicorp/terraform/config"
 	"github.com/hashicorp/terraform/config/module"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/pulumi/pulumi/pkg/util/contract"
 	"github.com/pulumi/tf2pulumi/gen"
 	"github.com/pulumi/tf2pulumi/il"
 )
@@ -102,198 +99,16 @@ func TestLowerToLiteral(t *testing.T) {
 	assert.Equal(t, "{\n    key: `module: foo/bar root: .`,\n}", computed)
 }
 
+func readFile(t *testing.T, path string) string {
+	bytes, err := ioutil.ReadFile(path)
+	if err != nil {
+		t.Fatalf("could not read file %s: %v", path, err)
+	}
+	return string(bytes)
+}
+
 func TestComments(t *testing.T) {
-	const hclText = `
-# Accept the AWS region as input.
-variable "aws_region" {
-	# Default to us-west-2
-	default = "us-west-2"
-}
-
-/*
-Specify provider details
-*/
-provider "aws" {
-	# Pull the region from a variable
-    region = "${var.aws_region}"
-}
-
-# Create a VPC.
-#
-# Note that the VPC has been tagged appropriately.
-resource "aws_vpc" "default" {
-    cidr_block = "10.0.0.0/16"  # Just one CIDR block
-	enable_dns_hostnames = true # Definitely want DNS hostnames.
-
-	# The tag collection for this VPC.
-	tags {
-		# Ensure that we tag this VPC with a Name.
-		Name = "test"
-	}
-}
-
-# Use some data sources.
-data "aws_subnet_ids" "default" {
-	vpc_id = "${aws_vpc.default.id}"
-}
-
-data "aws_availability_zones" "default" {}
-
-data "aws_availability_zone" "default" {
-	count = "${length(data.aws_availability_zones.default.ids)}"
-	zone_id = "${data.aws_availability_zones.default.zone_ids[count.index]}"
-}
-
-locals {
-	# The VPC details
-	vpc = {
-		# The ID
-		id = "${aws_vpc.default.id}"
-	}
-
-	# The region, again
-	region = "${var.aws_region}" // why not
-}
-
-// Create a security group.
-//
-// This group should allow SSH and HTTP access.
-resource "aws_security_group" "default" {
-	vpc_id = "${local.vpc["id"]}"
-
-	// SSH access from anywhere
-	ingress {
-		from_port   = 22
-		to_port     = 22
-		protocol    = "tcp"
-		// "0.0.0.0/0" is anywhere
-		cidr_blocks = ["0.0.0.0/0"]
-	}
-
-	// HTTP access from anywhere
-	ingress {
-		from_port   = 80
-		to_port     = 80
-		protocol    = "tcp" /* HTTP is TCP-only */
-		cidr_blocks = ["0.0.0.0/0"]
-	}
-
-	// outbound internet access
-	egress {
-		from_port   = 0
-		to_port     = 0
-		protocol    = "-1" // All
-		cidr_blocks = ["0.0.0.0/0"]
-	}
-
-	tags {
-		Vpc = "VPC ${var.aws_region}:${aws_vpc.default.id}"
-	}
-}
-
-/**
- * Output the SG name.
- *
- * We pull the name from the default SG.
- */
-output "security_group_name" {
-	/* Take the value from the default SG. */
-	value = "${aws_security_group.default.name}" # Neat!
-}
-`
-
-	const expectedText16 = `import * as pulumi from "@pulumi/pulumi";
-import * as aws from "@pulumi/aws";
-
-const config = new pulumi.Config();
-// Accept the AWS region as input.
-const awsRegion = config.get("awsRegion") || "us-west-2";
-
-// Create a VPC.
-//
-// Note that the VPC has been tagged appropriately.
-const defaultVpc = new aws.ec2.Vpc("default", {
-    cidrBlock: "10.0.0.0/16", // Just one CIDR block
-    enableDnsHostnames: true, // Definitely want DNS hostnames.
-    // The tag collection for this VPC.
-    tags: {
-        // Ensure that we tag this VPC with a Name.
-        Name: "test",
-    },
-});
-const defaultAvailabilityZones = pulumi.output(aws.getAvailabilityZones({}));
-// The region, again
-const region = awsRegion; // why not
-// The VPC details
-const vpc = [{
-    // The ID
-    id: defaultVpc.id,
-}];
-// Create a security group.
-//
-// This group should allow SSH and HTTP access.
-const defaultSecurityGroup = new aws.ec2.SecurityGroup("default", {
-    // outbound internet access
-    egress: [{
-        cidrBlocks: ["0.0.0.0/0"],
-        fromPort: 0,
-        protocol: "-1", // All
-        toPort: 0,
-    }],
-    ingress: [
-        // SSH access from anywhere
-        {
-            // "0.0.0.0/0" is anywhere
-            cidrBlocks: ["0.0.0.0/0"],
-            fromPort: 22,
-            protocol: "tcp",
-            toPort: 22,
-        },
-        // HTTP access from anywhere
-        {
-            cidrBlocks: ["0.0.0.0/0"],
-            fromPort: 80,
-            protocol: "tcp", // HTTP is TCP-only
-            toPort: 80,
-        },
-    ],
-    tags: {
-        Vpc: defaultVpc.id.apply(id => ` + "`VPC ${awsRegion}:${id}`" + `),
-    },
-    vpcId: vpc["id"],
-});
-const defaultAvailabilityZone: pulumi.Output<aws.GetAvailabilityZoneResult>[] = [];
-for (let i = 0; i < defaultAvailabilityZones.apply(defaultAvailabilityZones => defaultAvailabilityZones.ids.length); i++) {
-    defaultAvailabilityZone.push(defaultAvailabilityZones.apply(defaultAvailabilityZones => aws.getAvailabilityZone({
-        zoneId: defaultAvailabilityZones.zoneIds[i],
-    })));
-}
-// Use some data sources.
-const defaultSubnetIds = defaultVpc.id.apply(id => aws.ec2.getSubnetIds({
-    vpcId: id,
-}));
-
-// Output the SG name.
-//
-// We pull the name from the default SG.
-// Take the value from the default SG.
-export const securityGroupName = defaultSecurityGroup.name; // Neat!
-`
-
-	dir, err := ioutil.TempDir("", "")
-	if err != nil {
-		t.Fatalf("could not create temporary directory: %v", err)
-	}
-	defer func() {
-		contract.IgnoreError(os.RemoveAll(dir))
-	}()
-
-	err = ioutil.WriteFile(path.Join(dir, "main.tf"), []byte(hclText), 0600)
-	if err != nil {
-		t.Fatalf("could not create main.tf: %v", err)
-	}
-
-	conf, err := config.LoadDir(dir)
+	conf, err := config.LoadDir("testdata/test_comments")
 	if err != nil {
 		t.Fatalf("could not load config: %v", err)
 	}
@@ -313,85 +128,9 @@ export const securityGroupName = defaultSecurityGroup.name; // Neat!
 	err = gen.Generate([]*il.Graph{g}, lang)
 	assert.NoError(t, err)
 
+	expectedText16 := readFile(t, "testdata/test_comments/index.16.ts")
 	assert.Equal(t, expectedText16, b.String())
 
-	const expectedText17 = `import * as pulumi from "@pulumi/pulumi";
-import * as aws from "@pulumi/aws";
-
-const config = new pulumi.Config();
-// Accept the AWS region as input.
-const awsRegion = config.get("awsRegion") || "us-west-2";
-
-// Create a VPC.
-//
-// Note that the VPC has been tagged appropriately.
-const defaultVpc = new aws.ec2.Vpc("default", {
-    cidrBlock: "10.0.0.0/16", // Just one CIDR block
-    enableDnsHostnames: true, // Definitely want DNS hostnames.
-    // The tag collection for this VPC.
-    tags: {
-        // Ensure that we tag this VPC with a Name.
-        Name: "test",
-    },
-});
-const defaultAvailabilityZones = pulumi.output(aws.getAvailabilityZones({}));
-// The region, again
-const region = awsRegion; // why not
-// The VPC details
-const vpc = [{
-    // The ID
-    id: defaultVpc.id,
-}];
-// Create a security group.
-//
-// This group should allow SSH and HTTP access.
-const defaultSecurityGroup = new aws.ec2.SecurityGroup("default", {
-    // outbound internet access
-    egress: [{
-        cidrBlocks: ["0.0.0.0/0"],
-        fromPort: 0,
-        protocol: "-1", // All
-        toPort: 0,
-    }],
-    ingress: [
-        // SSH access from anywhere
-        {
-            // "0.0.0.0/0" is anywhere
-            cidrBlocks: ["0.0.0.0/0"],
-            fromPort: 22,
-            protocol: "tcp",
-            toPort: 22,
-        },
-        // HTTP access from anywhere
-        {
-            cidrBlocks: ["0.0.0.0/0"],
-            fromPort: 80,
-            protocol: "tcp", // HTTP is TCP-only
-            toPort: 80,
-        },
-    ],
-    tags: {
-        Vpc: pulumi.interpolate` + "`VPC ${awsRegion}:${defaultVpc.id}`" + `,
-    },
-    vpcId: vpc["id"],
-});
-const defaultAvailabilityZone: pulumi.Output<aws.GetAvailabilityZoneResult>[] = [];
-for (let i = 0; i < defaultAvailabilityZones.apply(defaultAvailabilityZones => defaultAvailabilityZones.ids.length); i++) {
-    defaultAvailabilityZone.push(defaultAvailabilityZones.apply(defaultAvailabilityZones => aws.getAvailabilityZone({
-        zoneId: defaultAvailabilityZones.zoneIds[i],
-    })));
-}
-// Use some data sources.
-const defaultSubnetIds = defaultVpc.id.apply(id => aws.ec2.getSubnetIds({
-    vpcId: id,
-}));
-
-// Output the SG name.
-//
-// We pull the name from the default SG.
-// Take the value from the default SG.
-export const securityGroupName = defaultSecurityGroup.name; // Neat!
-`
 	g, err = il.BuildGraph(module.NewTree("main", conf), &il.BuildOptions{
 		AllowMissingProviders: true,
 		AllowMissingVariables: true,
@@ -407,85 +146,9 @@ export const securityGroupName = defaultSecurityGroup.name; // Neat!
 	err = gen.Generate([]*il.Graph{g}, lang)
 	assert.NoError(t, err)
 
+	expectedText17 := readFile(t, "testdata/test_comments/index.17.ts")
 	assert.Equal(t, expectedText17, b.String())
 
-	const expectedText17PromptDataSources = `import * as pulumi from "@pulumi/pulumi";
-import * as aws from "@pulumi/aws";
-
-const config = new pulumi.Config();
-// Accept the AWS region as input.
-const awsRegion = config.get("awsRegion") || "us-west-2";
-
-// Create a VPC.
-//
-// Note that the VPC has been tagged appropriately.
-const defaultVpc = new aws.ec2.Vpc("default", {
-    cidrBlock: "10.0.0.0/16", // Just one CIDR block
-    enableDnsHostnames: true, // Definitely want DNS hostnames.
-    // The tag collection for this VPC.
-    tags: {
-        // Ensure that we tag this VPC with a Name.
-        Name: "test",
-    },
-});
-const defaultAvailabilityZones = aws.getAvailabilityZones({});
-// The region, again
-const region = awsRegion; // why not
-// The VPC details
-const vpc = [{
-    // The ID
-    id: defaultVpc.id,
-}];
-// Create a security group.
-//
-// This group should allow SSH and HTTP access.
-const defaultSecurityGroup = new aws.ec2.SecurityGroup("default", {
-    // outbound internet access
-    egress: [{
-        cidrBlocks: ["0.0.0.0/0"],
-        fromPort: 0,
-        protocol: "-1", // All
-        toPort: 0,
-    }],
-    ingress: [
-        // SSH access from anywhere
-        {
-            // "0.0.0.0/0" is anywhere
-            cidrBlocks: ["0.0.0.0/0"],
-            fromPort: 22,
-            protocol: "tcp",
-            toPort: 22,
-        },
-        // HTTP access from anywhere
-        {
-            cidrBlocks: ["0.0.0.0/0"],
-            fromPort: 80,
-            protocol: "tcp", // HTTP is TCP-only
-            toPort: 80,
-        },
-    ],
-    tags: {
-        Vpc: pulumi.interpolate` + "`VPC ${awsRegion}:${defaultVpc.id}`" + `,
-    },
-    vpcId: vpc["id"],
-});
-const defaultAvailabilityZone: aws.GetAvailabilityZoneResult[] = [];
-for (let i = 0; i < defaultAvailabilityZones.ids.length; i++) {
-    defaultAvailabilityZone.push(aws.getAvailabilityZone({
-        zoneId: defaultAvailabilityZones.zoneIds[i],
-    }));
-}
-// Use some data sources.
-const defaultSubnetIds = defaultVpc.id.apply(id => aws.ec2.getSubnetIds({
-    vpcId: id,
-}));
-
-// Output the SG name.
-//
-// We pull the name from the default SG.
-// Take the value from the default SG.
-export const securityGroupName = defaultSecurityGroup.name; // Neat!
-`
 	g, err = il.BuildGraph(module.NewTree("main", conf), &il.BuildOptions{
 		AllowMissingProviders: true,
 		AllowMissingVariables: true,
@@ -501,6 +164,6 @@ export const securityGroupName = defaultSecurityGroup.name; // Neat!
 	err = gen.Generate([]*il.Graph{g}, lang)
 	assert.NoError(t, err)
 
+	expectedText17PromptDataSources := readFile(t, "testdata/test_comments/index.v1.ts")
 	assert.Equal(t, expectedText17PromptDataSources, b.String())
-
 }
